jpegxl/interop-2026/features/vardct-large-blocks.html

<!DOCTYPE html>
<title>JPEG XL interop: VarDCT larger image stress</title>
<link rel="help" href="https://github.com/web-platform-tests/interop-jpegxl">
<meta name="assert" content="Decodes larger VarDCT fixture.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../resources/interop-helpers.js"></script>
<script>

promise_test(async () => {
  const img = await loadImage('../resources/green_queen_vardct_e3.jxl');
  assert_equals(img.naturalWidth, 438);
  assert_equals(img.naturalHeight, 589);
}, 'Larger VarDCT fixture decodes to expected dimensions.');

</script>
